requirements.txt
click==8.1.7
openai>=1.12.0
gitpython==3.1.43
python-dotenv==1.0.1